# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/polls": {
      "get": {
        "description": "Returns polls, optionally filtered by event or status. Accessible by client, eventmanager, osc, attendee, and superuser.",
        "operationId": "PollsController_findAll",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter by event ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by poll status",
            "schema": {
              "enum": [
                "open",
                "closed"
              ],
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "superuser | eventmanager | client | osc | attendee",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Polls list returned"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get all polls",
        "tags": [
          "Polls"
        ]
      },
      "post": {
        "description": "Client or eventmanager creates a poll for an event. Requires at least 2 options.",
        "operationId": "PollsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | client | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "client"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePollDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Poll created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Create a new poll",
        "tags": [
          "Polls"
        ]
      }
    },
    "/polls/{id}": {
      "get": {
        "operationId": "PollsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Poll ID (e.g. poll1)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "superuser | eventmanager | client | osc | attendee",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll details"
          },
          "404": {
            "description": "Poll not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get poll by ID",
        "tags": [
          "Polls"
        ]
      },
      "patch": {
        "description": "Client or eventmanager can update a poll. Updating options resets vote counts.",
        "operationId": "PollsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Poll ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | client | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "client"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePollDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Poll updated"
          },
          "404": {
            "description": "Poll not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Update poll (question, options, or status)",
        "tags": [
          "Polls"
        ]
      },
      "delete": {
        "description": "Client or eventmanager permanently removes a poll.",
        "operationId": "PollsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Poll ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | client | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll deleted"
          },
          "404": {
            "description": "Poll not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete a poll",
        "tags": [
          "Polls"
        ]
      }
    },
    "/polls/{id}/results": {
      "get": {
        "description": "Returns vote counts and percentage breakdown per option.",
        "operationId": "PollsController_getResults",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Poll ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "superuser | eventmanager | client | osc | attendee",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll results with percentages"
          },
          "404": {
            "description": "Poll not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get poll results with vote percentages",
        "tags": [
          "Polls"
        ]
      }
    },
    "/polls/{id}/vote": {
      "post": {
        "description": "Attendee votes on an open poll. Each attendee (by email) can vote only once per poll.",
        "operationId": "PollsController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Poll ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: attendee",
            "required": true,
            "schema": {
              "type": "string",
              "example": "attendee"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VotePollDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vote registered successfully"
          },
          "400": {
            "description": "Poll is closed or invalid option"
          },
          "409": {
            "description": "Already voted"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Vote on a poll",
        "tags": [
          "Polls"
        ]
      }
    },
    "/polls/{id}/close": {
      "patch": {
        "description": "Client or eventmanager closes the poll.",
        "operationId": "PollsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Poll ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | client | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll closed"
          },
          "404": {
            "description": "Poll not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Close a poll (stop accepting votes)",
        "tags": [
          "Polls"
        ]
      }
    },
...
  },
  "info": {
    "title": "WEvents API",
...
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role (superuser | eventmanager | client | osc | attendee)"
      }
    },
    "schemas": {
...
      "CreatePollDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "example": "e1",
            "description": "Event ID this poll belongs to"
          },
          "question": {
            "type": "string",
            "example": "Which session did you enjoy most?",
            "description": "The poll question"
          },
          "options": {
            "example": [
              "Keynote",
              "Workshop A",
              "Workshop B"
            ],
            "description": "List of answer options (min 2)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdBy": {
            "type": "string",
            "example": "c1",
            "description": "Client ID who created the poll"
          }
        },
        "required": [
          "eventId",
          "question",
          "options"
        ]
      },
      "VotePollDto": {
        "type": "object",
        "properties": {
          "option": {
            "type": "string",
            "example": "Workshop A",
            "description": "The option the attendee is voting for"
          },
          "voterEmail": {
            "type": "string",
            "example": "rahul.s@gmail.com",
            "description": "Voter email (attendee)"
          }
        },
        "required": [
          "option",
          "voterEmail"
        ]
      },
      "UpdatePollDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "example": "Updated question text?",
            "description": "Update the poll question"
          },
          "options": {
            "example": [
              "Option A",
              "Option B",
              "Option C"
            ],
            "description": "Update poll options",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "example": "open",
            "enum": [
              "open",
              "closed"
            ],
            "description": "Poll status"
          }
        }
      },
...
    }
  }
}